Modification of algorithm and methods, improved readability
Removed the product state cost to improve the convergence and accuracy of the code. The reconstructed states now have normalised product states multiplied by the coefficients. Hence, the changes were made in the calculation of the projective norm. The coefficient cost now calculates the square of the number of terms greater than a threshold to effectively reduce the coefficients. Removed extra comments and helper print lines to improve the readability.